Isolate PMA automation unified mirror
