+++ tests/core/test_pma_automation_store.py
@@ -7,10 +7,12 @@ import pytest
 from codex_autorunner.core.automation import AutomationStore
 from codex_autorunner.core.managed_thread_store import ManagedThreadStore
 from codex_autorunner.core.orchestration import OrchestrationBindingStore
+from codex_autorunner.core.pma_automation_mirror import PmaAutomationMirror
 from codex_autorunner.core.pma_automation_store import (
     PmaAutomationStore,
     PmaAutomationThreadNotFoundError,
 )
+from codex_autorunner.core.pma_automation_unified import PmaUnifiedAutomationAdapter
 from codex_autorunner.core.pma_domain.models import PmaDispatchDecision
 
 
@@ -122,6 +124,60 @@ def test_timer_create_touch_and_cancel_mirror_unified_schedule(tmp_path) -> None
     cancelled = automation_store.get_schedule(f"pma-timer:{timer_id}")
     assert cancelled.state == "cancelled"
     assert cancelled.next_fire_at is None
+
+
+def test_subscription_mirror_returns_structured_success(tmp_path) -> None:
+    store = PmaAutomationStore(tmp_path)
+    subscription, deduped = store.upsert_subscription(
+        event_types=["flow_failed"],
+        repo_id="repo-1",
+        lane_id="pma:default",
+    )
+
+    assert deduped is False
+    result = PmaAutomationMirror(tmp_path).mirror_subscription_rule(subscription)
+
+    assert result.ok is True
+    assert result.operation == "mirror_subscription_rule"
+    assert result.identifier == subscription.subscription_id
+    assert result.rule_id == f"builtin:pma:subscription:{subscription.subscription_id}"
+    assert AutomationStore(tmp_path).get_rule(result.rule_id) is not None
+
+
+def test_timer_mirror_failure_returns_structured_error_and_logs(
+    tmp_path, monkeypatch, caplog: pytest.LogCaptureFixture
+) -> None:
+    store = PmaAutomationStore(tmp_path)
+    timer, deduped = store.upsert_timer(
+        due_at="2026-01-01T00:00:00Z",
+        thread_id="thread-123",
+    )
+    assert deduped is False
+
+    def _raise_runtime_error(self, *, timer):
+        _ = self, timer
+        raise RuntimeError("unified store unavailable")
+
+    monkeypatch.setattr(
+        PmaUnifiedAutomationAdapter,
+        "mirror_timer_schedule",
+        _raise_runtime_error,
+    )
+
+    with caplog.at_level(
+        logging.ERROR, logger="codex_autorunner.core.pma_automation_mirror"
+    ):
+        result = PmaAutomationMirror(tmp_path).mirror_timer_schedule(timer)
+
+    assert result.ok is False
+    assert result.operation == "mirror_timer_schedule"
+    assert result.identifier == timer.timer_id
+    assert "unified store unavailable" in result.error
+    assert any(
+        "Failed to mirror PMA timer into unified automation schedule"
+        in record.getMessage()
+        for record in caplog.records
+    )
 
 
 def test_watchdog_timer_refires_until_touched(tmp_path) -> None:
